requirements.txt
praw==7.7.1
pandas==2.0.3
python-dotenv==1.0.0
tqdm==4.66.1
datetime
